+++ Source/Main Dialogs/CommandCenter/Scripting/OrcaScript.h
@@ -1,4 +1,4 @@
-extern int OrcaScriptYYINPUT(char* theBuffer,int maxSize);
+extern int OrcaScriptYYINPUT(char* theBuffer,unsigned long maxSize);
 #undef YY_INPUT
 #define YY_INPUT(b,r,s) (r = OrcaScriptYYINPUT(b,s))
